chore: move components directory to root and update all imports
