+++ __tests__/useApi.test.ts
@@ -20,7 +20,7 @@ import { mocks } from "./setup";
 import { expect, describe, it, beforeEach } from "bun:test";
 import { renderHook, waitFor } from "@testing-library/react-native";
 import { act } from "react";
-import useApi from "../app/components/useApi";
+import useApi from "../components/useApi";
 
 describe("useApi", () => {
 	beforeEach(() => {
